feat(tui): remember the request card tab across region moves
